+++ src/app.module.ts
@@ -3,25 +3,10 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { AuthorsModule } from './authors/authors.module';
 import { PostsModule } from './posts/posts.module';
-// import { TypeOrmModule } from '@nestjs/typeorm';
-// import { Author } from './authors/entities/author.entity';
-// import { Post } from './posts/entities/post.entity';
+import { DatabaseModule } from './database/database.module';
 
 @Module({
-  imports: [
-    // TypeOrmModule.forRoot({
-    //   type: 'postgres',
-    //   host: 'localhost',
-    //   port: 3306,
-    //   username: 'root',
-    //   password: 'root',
-    //   database: 'test',
-    //   entities: [Author, Post],
-    //   synchronize: true,
-    // }),
-    AuthorsModule,
-    PostsModule,
-  ],
+  imports: [AuthorsModule, PostsModule, DatabaseModule],
   controllers: [AppController],
   providers: [AppService],
 })