+++ src/database/database.module.ts
@@ -1,0 +1,29 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Author } from 'src/authors/entities/author.entity';
+import { Post } from 'src/posts/entities/post.entity';
+import { DataSourceOptions } from 'typeorm';
+
+export const dataSourceOptions: DataSourceOptions = {
+  type: 'postgres',
+  host: 'db',
+  port: 5432,
+  username: 'postgres',
+  password: 'docker',
+  database: 'personal-blog',
+  entities: [Author, Post],
+  synchronize: true,
+};
+
+@Module({
+  imports: [
+    TypeOrmModule.forRootAsync({
+      useFactory: async () => {
+        return {
+          ...dataSourceOptions,
+        };
+      },
+    }),
+  ],
+})
+export class DatabaseModule {}