+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+### v0.3.0 (27/10/2024)
+- Adds postgres connection
+- Adjustments in authors modules
+
 ### v0.2.0 (01/10/2024)
 - Adds docker integration